Delete gem 'asset_sync' as it is not being used
Rename carousel.js to carousel.es6 in /assets/...

This should be fine for pushing to heroku now. It
had been breaking on the 'lets' in the carousel 
file and upon not having a fog file for 'ASW'.

I might want the gem 'asset_sync' and 'fog' later.
Need to do more research, but for now seems 
unnecessary.